+++ CHANGELOG.md
@@ -76,6 +76,7 @@ All notable changes to this project will be documented in this file.
 - Made UI to add new rows to _Table_ attribute clearer
 - Attribute type backend logic moved to separate classes
 - Set default value in parent element in _Add Attribute_ modal
+- Rework System/User Preferences and allow plugins to add their own preferences
 - Values in Dropdown attributes are now sorted alphabetically
 - Values in Dropdown attributes are now searchable
 - Font size in dropdown menus